Readd instrumentation.
Figured out that if one doesn't add a file to the list
that file won't have it added. I thought that it had to
be on that list to be considered. I will need to add a
ban-files list, because the only option right now is
to restrict-to-files, which is the opposite of what I want.